Fix Home not reloading after adding a new Wallet by changing it to an activity
