+++ app/src/main/AndroidManifest.xml
@@ -37,6 +37,10 @@
             android:exported="true"
             android:screenOrientation="portrait">
         </activity>
+        <activity
+            android:name=".ui.dialogs.AddWallet"
+            android:exported="true">
+        </activity>
     </application>
 
 </manifest>
