+++ app/src/main/res/layout/activity_add_payment.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:theme="@style/Theme.FAP.Add_Payment"
-    tools:context=".ui.login.Login">
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
